# le-taxi-api-node.js/swagger.fr.json
{
    "info": {
        "version": "1.0",
        "title": "API version 2.0"
    },
    "paths": {
...
        "/hails/{hail_id}/": {
            "put": {
                "tags": [ "hails" ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": { "$ref": "#/definitions/hail_Hail" }
                    }
                },
                "parameters": [
                    {
                        "required": true,
                        "type": "string",
                        "name": "hail_id",
                        "in": "path"
                    },
                    {
                        "required": true,
                        "in": "body",
                        "name": "payload",
                        "schema": { "$ref": "#/definitions/hail_expect_put" }
                    },
                    {
                        "description": "An optionnal fields mask",
                        "format": "mask",
                        "type": "string",
                        "name": "X-Fields",
                        "in": "header"
                    }
                ],
                "operationId": "put_hail_id"
            },
            "get": {
                "tags": [ "hails" ],
                "responses": { "200": { "description": "Success" } },
                "parameters": [
                    {
                        "required": true,
                        "type": "string",
                        "name": "hail_id",
                        "in": "path"
                    }
                ],
                "operationId": "get_hail_id"
            }
        },
...
        "/hails/": {
            "post": {
                "tags": [ "hails" ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "schema": { "$ref": "#/definitions/hail_Hail" }
                    }
                },
                "parameters": [
                    {
                        "required": true,
                        "in": "body",
                        "name": "payload",
                        "schema": { "$ref": "#/definitions/hail_expect_post" }
                    }
                ],
                "operationId": "post_hail"
            }
        },
...
    },
...
    "basePath": "/",
    "produces": [ "application/json", "text/html" ],
    "host": "{{ host }}",
    "definitions": {
...
        "hail_expect_put_details": {
            "properties": {
                "status": {
                    "enum": [ "emitted", "received", "sent_to_operator", "received_by_operator", "received_by_taxi", "accepted_by_taxi", "accepted_by_customer", "declined_by_taxi", "declined_by_customer", "incident_customer", "incident_taxi", "timeout_customer", "timeout_taxi", "outdated_customer", "outdated_taxi", "failure" ],
                    "type": "string",
                    "description": "status",
                    "example": "emitted"
                },
                "incident_customer_reason": {
                    "enum": [ "", "mud_river", "parade", "earthquake" ],
                    "type": "string",
                    "description": "incident_customer_reason",
                    "example": ""
                },
                "customer_lat": {
                    "type": "number",
                    "description": "customer_lat"
                },
                "rating_ride": {
                    "type": "integer",
                    "description": "rating_ride"
                },
                "rating_ride_reason": {
                    "enum": [ "ko", "payment", "courtesy", "route", "cleanliness", "late", "no_credit_card", "bad_itinerary", "dirty_taxi" ],
                    "type": "string",
                    "description": "rating_ride_reason",
                    "example": "ko"
                },
                "customer_lon": {
                    "type": "number",
                    "description": "customer_lon"
                },
                "incident_taxi_reason": {
                    "enum": [ "no_show", "address", "traffic", "breakdown", "traffic_jam", "garbage_truck" ],
                    "type": "string",
                    "description": "incident_taxi_reason",
                    "example": "no_show"
                },
                "customer_address": {
                    "type": "string",
                    "description": "customer_address"
                },
                "customer_phone_number": {
                    "type": "string",
                    "description": "customer_phone_number"
                },
                "reporting_customer": {
                    "type": "boolean",
                    "description": "reporting_customer"
                },
                "reporting_customer_reason": {
                    "enum": [ "ko", "payment", "courtesy", "route", "cleanliness", "late", "aggressive", "no_show" ],
                    "type": "string",
                    "description": "reporting_customer_reason",
                    "example": "ko"
                }
            }
        },
...
        "coordinates_descriptor": {
            "properties": {
                "lat": { "type": "number" },
                "lon": { "type": "number" }
            }
        },
...
        "hail_expect_put": {
            "properties": {
                "data": {
                    "items": { "$ref": "#/definitions/hail_expect_put_details" },
                    "type": "array",
                    "maxItems": 1
                }
            }
        },
...
        "hail_expect_post": {
            "properties": {
                "data": {
                    "items": { "$ref": "#/definitions/hail_expect_post_details" },
                    "type": "array",
                    "maxItems": 1
                }
            }
        },
        "hail_Hail_details": {
            "required": [ "creation_datetime", "customer_address", "customer_id", "customer_lat", "customer_lon", "customer_phone_number", "operateur", "status" ],
            "properties": {
                "status": {
                    "enum": [ "emitted", "received", "sent_to_operator", "received_by_operator", "received_by_taxi", "accepted_by_taxi", "accepted_by_customer", "declined_by_taxi", "declined_by_customer", "incident_customer", "incident_taxi", "timeout_customer", "timeout_taxi", "outdated_customer", "outdated_taxi", "failure" ],
                    "type": "string",
                    "description": "status",
                    "example": "emitted"
                },
                "taxi": { "$ref": "#/definitions/hail_taxi" },
                "creation_datetime": {
                    "type": "string",
                    "description": "creation_datetime",
                    "format": "date-time"
                },
                "customer_lat": {
                    "type": "number",
                    "description": "customer_lat"
                },
                "rating_ride": {
                    "type": "integer",
                    "description": "rating_ride"
                },
                "rating_ride_reason": {
                    "enum": [ "ko", "payment", "courtesy", "route", "cleanliness", "late", "no_credit_card", "bad_itinerary", "dirty_taxi" ],
                    "type": "string",
                    "description": "rating_ride_reason",
                    "example": "ko"
                },
                "customer_lon": {
                    "type": "number",
                    "description": "customer_lon"
                },
                "reporting_customer_reason": {
                    "enum": [ "ko", "payment", "courtesy", "route", "cleanliness", "late", "aggressive", "no_show" ],
                    "type": "string",
                    "description": "reporting_customer_reason",
                    "example": "ko"
                },
                "incident_taxi_reason": {
                    "enum": [ "no_show", "address", "traffic", "breakdown", "traffic_jam", "garbage_truck" ],
                    "type": "string",
                    "description": "incident_taxi_reason",
                    "example": "no_show"
                },
                "customer_address": {
                    "type": "string",
                    "description": "customer_address"
                },
                "last_status_change": {
                    "type": "string",
                    "description": "last_status_change",
                    "format": "date-time"
                },
                "customer_phone_number": {
                    "type": "string",
                    "description": "customer_phone_number"
                },
                "taxi_phone_number": {
                    "type": "string",
                    "description": "taxi_phone_number"
                },
                "reporting_customer": {
                    "type": "boolean",
                    "description": "reporting_customer"
                },
                "operateur": { "type": "string" },
                "incident_customer_reason": {
                    "enum": [ "", "mud_river", "parade", "earthquake" ],
                    "type": "string",
                    "description": "incident_customer_reason",
                    "example": ""
                },
                "customer_id": {
                    "type": "string",
                    "description": "customer_id"
                },
                "id": { "type": "string" }
            }
        },
        "hail_expect_post_details": {
            "required": [ "customer_address", "customer_id", "customer_lat", "customer_lon", "customer_phone_number", "operateur", "taxi_id" ],
            "properties": {
                "customer_lat": {
                    "type": "number",
                    "description": "customer_lat"
                },
                "customer_lon": {
                    "type": "number",
                    "description": "customer_lon"
                },
                "customer_address": {
                    "type": "string",
                    "description": "customer_address"
                },
                "taxi_id": { "type": "string" },
                "customer_phone_number": {
                    "type": "string",
                    "description": "customer_phone_number"
                },
                "operateur": { "type": "string" },
                "customer_id": {
                    "type": "string",
                    "description": "customer_id"
                }
            }
        },
...
        "hail_Hail": {
            "properties": {
                "data": {
                    "items": { "$ref": "#/definitions/hail_Hail_details" },
                    "type": "array"
                }
            }
        },
...
        "hail_taxi": {
            "properties": {
                "position": { "$ref": "#/definitions/coordinates_descriptor" },
                "id": { "type": "string" },
                "last_update": { "type": "integer" }
            }
        }
    },
    "swagger": "2.0",
    "consumes": [ "application/json" ]
}